Sync user in container with local
Locally, files are owned by some user with ID 1000 (name irrelevant). This ownership persists for mounted files inside the container.
Nginx, which starts PHP, runs under the username and group `www-data`. This user/group ID is probably something other than 1000.
The permissions on the mounted files inside the container allow everyone to read, but only the owner to write.

This change sets the ID of the user/group `www-data` inside the container to 1000, effectively making `www-data` the owner of the files without changing any file attributes.