Fix uploads directory not being writable
This prevented product images from being added to media library.
But it also prevented logging, so it was hard to debug.
It was due to the `wp-content` directory being owned by `root`, because files were being mounted into `wp-content/plugins` before the server had the chance to set up WP.
Could not `chown` the whole `wp-content` directory because the mounted files cause FS recursion.